fix: 방 참가 등록을 인식하지 못하는 버그 수정
Reference 리팩토링
