+++ Discussion-Korea/FirebasePlatform/Reference/GuideReference.swift
@@ -1,0 +1,13 @@
+//
+//  GuideReference.swift
+//  Discussion-Korea
+//
+//  Created by 이청수 on 2022/08/31.
+//
+
+import FirebaseDatabase
+import RxSwift
+
+final class GuideReference {
+
+}
