+++ Discussion-Korea/FirebasePlatform/Reference/LawReference.swift
@@ -1,0 +1,12 @@
+//
+//  LawReference.swift
+//  Discussion-Korea
+//
+//  Created by 이청수 on 2022/08/31.
+//
+
+import RxSwift
+
+final class LawReference {
+
+}
